requirements.txt
streamlit
pdfplumber
pandas
openpyxl
thefuzz
python-Levenshtein
fpdf2
